+++ activestorage/lib/active_storage/direct_upload.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module ActiveStorage
+  class DirectUpload
+    def initialize(blob)
+      @blob = blob
+
+      @method = @blob.service_method_for_direct_upload
+      @url = @blob.service_url_for_direct_upload
+      @headers = @blob.service_headers_for_direct_upload
+    end
+
+    def as_json
+      { method: @method, url: @url, headers: @headers }
+    end
+  end
+end